Update to v0.2.6
- Fix entity parser encoding all input, to only encoding descriptions.
- Fix output parsing showing entity codes rather than the symbols they represent.
- Fix Page Settings checkbox for enabling Slider checking wrong value.
- Fix broken thumbnail on bookings page.
- Form submission serializer not parsing correctly selected event/service to backend.
- Fix reference to incorrect variable in add_contantus.
- Fix styling for required form fields that wasn't diplaying correctly.
- Improve Forms for all client sites for better validation and requirement notifications.
- Update Summernote WYSIWYG Editor.
- Adjust logic to stop Recommended Retail Price from showing in items list, in favour of Cost Price.
- Update Summernote plugin summernote-cleaner.
- Fix Account Required to Purchase option not working correctly.
- Add Form feature that fetch's postcode information autofilling Suburb and State fields.
- Fix Accountnotification warning tag element not being parsed away correctly.
- Add Video Playlist editor for pages, and front end parsing and markup.
- Remove unused emoji images.
- Fix Schema Product Prices when no price is set, or is a text value, to not produce cost schema markup.

+++ core/purge.php
@@ -19,7 +19,7 @@ $tbl=isset($_POST['t'])?filter_input(INPUT_POST,'t',FILTER_SANITIZE_STRING):filt
 $col=isset($_POST['c'])?filter_input(INPUT_POST,'c',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'c',FILTER_SANITIZE_STRING);
 $uid=isset($_SESSION['uid'])?$_SESSION['uid']:0;
 $el='l_';
-if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'&&$tbl!='forumTopics'&&$tbl!='pageviews'){
+if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'&&$tbl!='forumTopics'&&$tbl!='pageviews'&&$tbl!='playlist'){
   $s=$db->prepare("SELECT * FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $s->execute([':id'=>$id]);
   $r=$s->fetch(PDO::FETCH_ASSOC);
@@ -77,7 +77,11 @@ if($id!=0&&$tbl=='seo'){
   $q->execute([':id'=>$id]);
   $el='l_';
 }
-if($id!=0&&$id!='activity'&&$tbl!='pageviews'){
+if($tbl=='playlist'){
+  $s=$db->prepare("DELETE FROM `".$prefix."choices` WHERE `id`=:id");
+  $s->execute([':id'=>$id]);
+}
+if($id!=0&&$id!='activity'&&$tbl!='pageviews'&&$tbl!='playlist'){
   $q=$db->prepare("DELETE FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $q->execute([':id'=>$id]);
   if($tbl=='media')$el='media_items_';
